make live tickets comp and whatsapp text fixed
